Added Group controller, service, event and response to request the list of saved cities on a single call. Refactored base architecture to use Activity and Fragment from non-support library to support usage of PreferenceFragment. Saved some default values on SharedPreferences using CacheManager. Implement caching expiry date Hide current weather when scrolling down on home screen (animation) Saving locations and toggle buttons Saving and displaying list of saved locations (recyclerView) Settings screen completed Saving of last zip code searched Change zip code search impl to String instead of Int Subscribe to some broadcast receiver

+++ app/src/main/java/com/allexis/weatherapp/core/persist/data/SavedLocation.java
@@ -38,7 +38,7 @@ public final class SavedLocation {
     }
 
     public static void removeSavedLocation(int cityId) {
-        SAVED_LOCATIONS.remove(cityId);
+        SAVED_LOCATIONS.remove((Integer) cityId);
         CacheManager.getInstance().setSavedLocations(TextUtils.join(TextUtil.SEPARATOR_COMMA, SAVED_LOCATIONS.toArray()));
     }
